pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "dl_cv_lectures"
version = "0.1.0"
description = "Deep Learning for Computer Vision Lecture Tools"
authors = [
    { name = "Marco Willi", email = "marco.willi@fhnw.ch" }
]


dependencies = [
    # --------- data & analysis -------- #
    "pandas",
    "numpy",
    # --------- visualize -------- #
    "matplotlib",
    "seaborn",
    # --------- data & image processing--------- #
    "Pillow",
    "opencv-python-headless",
    "scikit-image",
    "kornia",
    # --------- machine learning --------- #
    "mlxtend",
    "scikit-learn",
    "torch",
    "torchvision",
    "torchaudio",
    "torchmetrics",
    "torchshow",
    "torchviz",
    "torchinfo",
    "captum",
    "lightning",
    "timm",
    "pyppeteer",
    "transformers",
    "optuna",
    "ray[train, tune]",
    # --------- Experiment Tracking --------- #
    "tensorboard",
    "wandb",
    # --------- juypter --------- #
    "jupyter",
    "ipykernel",
    "ipywidgets",
    "xeus-python",
    "python-gitlab",
    "jupyter-black",
    # --------- Utilities --------- #
    "tqdm",
    "python-dotenv",
    "pyrootutils",
    "gdown"
]

[project.optional-dependencies]
dev = [
    "black",
    "argcomplete",
    "pytest",
    "mypy",
    "pre-commit",
    "rich",
    "ruff",
    "isort"
]
